split models on logical packages
