+++ src/test/java/executor/service/model/config/WebDriverConfigTest.java
@@ -1,5 +1,6 @@
-package executor.service.model;
+package executor.service.model.config;
 
+import executor.service.model.configs.WebDriverConfig;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
