+++ src/test/java/executor/service/model/model/ScenarioTest.java
@@ -1,5 +1,7 @@
-package executor.service.model;
+package executor.service.model.model;
 
+import executor.service.model.entity.Scenario;
+import executor.service.model.entity.Step;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
