+++ src/test/java/executor/service/model/proxy/ProxyConfigHolderTest.java
@@ -1,5 +1,8 @@
-package executor.service.model;
+package executor.service.model.proxy;
 
+import executor.service.model.proxy.ProxyConfigHolder;
+import executor.service.model.proxy.ProxyCredentials;
+import executor.service.model.proxy.ProxyNetworkConfig;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
